+++ lib/rgeo/feature/surface.rb
@@ -49,7 +49,7 @@ module RGeo
       # Returns a floating-point scalar value.
 
       def area
-        raise Error::UnsupportedOperation, "Method Surface#area not defined."
+        raise Error::UnsupportedOperation, "Method #{self.class.name}#area not defined."
       end
 
       # === SFS 1.1 Description
@@ -62,7 +62,7 @@ module RGeo
       # Returns an object that supports the Point interface.
 
       def centroid
-        raise Error::UnsupportedOperation, "Method Surface#centroid not defined."
+        raise Error::UnsupportedOperation, "Method #{self.class.name}#centroid not defined."
       end
 
       # === SFS 1.1 Description
@@ -74,7 +74,7 @@ module RGeo
       # Returns an object that supports the Point interface.
 
       def point_on_surface
-        raise Error::UnsupportedOperation, "Method Surface#point_on_surface not defined."
+        raise Error::UnsupportedOperation, "Method #{self.class.name}#point_on_surface not defined."
       end
     end
   end