+++ test/geos_ffi/multi_polygon_test.rb
@@ -13,8 +13,6 @@ class GeosFFIMultiPolygonTest < Minitest::Test # :nodoc:
 
   def create_factories
     @factory = RGeo::Geos.factory(native_interface: :ffi)
-    @lenient_factory = RGeo::Geos.factory(lenient_multi_polygon_assertions: true,
-                                          native_interface: :ffi)
   end
 
   # Centroid of an empty should return an empty collection